Learn to compute winning odds given a board
The program can now compute not only all-in preflop equities, but also
it can account for a flop, a flop and a turn, or even a full
board (flop, turn and river).

As one could expect, computing odds on a given board (even just a flop)
is way faster than computing pre-flop odds, given the reduced number
of combinations to test.

Add an example for this new feature in the README.md